fix: misc compilation fixes and warning cleanup
- config.rs: fix missing closing brace on default_context_window(),
  increase max_parallels 3 -> 8 for better LLM concurrency
- tests/common/mod.rs: add context_window field to test LLM config
- ollama_native.rs: fix context_window cast i32 -> u64 for ollama-rs API
- integrations/mod.rs: remove unused DocumentationManifest re-export
- llm/client/mod.rs: suppress dead_code warning on use_native_ollama()
- litho-cli/main.rs: remove unused import and unnecessary mut
- litho-qmd-storage/Cargo.toml: add native-tls and postgres-native-tls

+++ crates/litho-generator/src/llm/client/ollama_native.rs
@@ -135,7 +135,7 @@ impl OllamaNativeClient {
         // Context window – the most critical parameter.
         // Configurable via `context_window` in litho.toml [llm] section.
         // Gemma3 12B-IT-QAT supports up to 131072; default is 32768.
-        opts = opts.num_ctx(config.context_window as i32);
+        opts = opts.num_ctx(config.context_window as u64);
 
         // Generation limit.
         opts = opts.num_predict(config.max_tokens as i32);
